+++ plugin/pied-web-ux/index.php
@@ -4,7 +4,7 @@ class PiedWebUxPlugin extends \RainLoop\Plugins\AbstractPlugin
 {
     const NAME = 'Pied Web UX',
         AUTHOR = 'Pied Web',
-        VERSION = '1.8.6',
+        VERSION = '1.8.6.1',
         RELEASE = '2026-09-18',
         REQUIRED = '2.38.2',
         LICENSE = 'AGPL v3',